A new XCFramework
The NurAPI Bluetooth framework is now an XCFramework. It allows for different architectures to be more easily added to the framework and makes it easier to use in client code. Functionality is identical.

The demo project is updated to use the new framework. Note the removed “Run script” build stage.